requirements.txt
# AI智能陪伴系统依赖列表
Django>=5.2.1,<5.3
djangorestframework>=3.14.0,<3.15
python-dotenv>=1.0.0,<1.1
Pillow>=10.0.0,<10.1
requests>=2.31.0,<2.32
whitenoise>=6.5.0,<6.6
gunicorn>=21.2.0,<21.3
openai>=1.1.0,<1.2
